arm/mach-aspeed: Allow to disable WDT2
The IBM Genesis3 supports booting from second flash using WDT2, but
there's no working code to poke the WDT2 and it takes too long for the
kernel to load to poke the watchdog.

As it's an evaluation platform disable this feature for now.
Unselecting this Kconfig disables the WDT2 in early platform code and
prevents 2nd firmware from being launched during normal boot.

Signed-off-by: Patrick Rudolph <patrick.rudolph@9elements.com>